:wrench: Edit variable
Edit the variable name on 'Association with Risk Factors' tap from 'percent_increase_riskfactor' is 'riskfactor_percent_increase'.
According to the usability test, some users are unable to choose their preferred risk factor. Because on the screen of variable names, not all of them are visible, only the word front.